Assume the source PHP image is up-to-date
Instead of increasing the layer size by installing our own updates, assume that
the PHP image we use as a source has all of the important ones already, and
rebuild on top of that when important fixes are released instead of installing
it ourselves. This should reduce download and update size somewhat.